Run ChooseArch_windows.js to b as a before_build hook.
Update ChooseArch_windows.js to work of project root path.

Move ChooseArch_*.js files to scripts/ folder
