+++ scripts/ChooseArch_Win8.js
